Moved the Countdown to a new file
